[REFACTOR][OrcJIT] Isolate LLVM patches under llvm_patches/
Moves the two ObjectLinkingLayer plugins whose entire reason to exist is
an upstream LLVM defect out of orcjit_session.cc into their own files
under addons/tvm_ffi_orcjit/src/ffi/llvm_patches/:

  - gotpcrelx_fix.{h,cc}     (x86_64 Linux GOTPCRELX relaxation bug)
  - init_fini_plugin.{h,cc}  (ELFNixPlatform gap, COFFPlatform stalled)

Each file opens with a fixed-shape header (LLVM issue link, affected
versions, trigger, symptom, Removal checklist) so that when upstream
lands a fix we can delete the file and the matching include / plugin
registration mechanically. README.md indexes the patches and their
removal criteria.

The arena memory manager stays at the top level -- it is a design-level
feature (contiguous r-x layout, THP, faster teardown) that we keep even
after LLVM #173269 is fixed, not a workaround.

Also reduces the default arena capacity from 4 GB (x86_64) / 8 GB
(AArch64) to 1 GB on both architectures. 1 GB covers typical ML JIT
workloads while staying well under the PC-relative relocation limit
(x86_64 +/-2 GB, AArch64 +/-4 GB) and is friendlier to memory-constrained
hosts (containers, CI runners).

No logic changes: orcjit_session.cc shrinks from 864 -> 498 lines of
pure code motion. CI-covered build and tests pass on aarch64 Linux.
